WIN2030-15279:feat(console.c):Add devmem-r/w command
Changelogs:
1.Add devmem-r <addr in hex> to read som memory/io address through SPI
2.Add devmem-w <addr in hex> <value in hex> to write som memory/io through SPI
This is a backdoor for debugging the som

+++ Core/Inc/hf_common.h
@@ -227,7 +227,7 @@ extern UART_HandleTypeDef huart3;
 
 /* functions prototypes ---------------------------------------------*/
 void hexstr2mac(uint8_t *mac, const char *hexstr);
-uint32_t atoh(const char *in, uint32_t len);
+uint64_t atoh(const char *in, uint32_t len);
 
 void hf_http_task(void *argument);
 void es_eeprom_wp(uint8_t flag);
